refactored. added migrations for order and order items

+++ migrations/20230612055450_create_users_table.up.sql
@@ -1,13 +1,13 @@
 -- Add up migration script here
 CREATE TABLE `users` (
-  id int auto_increment not null,
+  id bigint unsigned auto_increment not null,
   name varchar(255) not null,
   email varchar(255) null,
   phone varchar(255) null,
   password_hash varchar(255) null,
   otp_secret varchar(255) null,
-  role int not null,
-  status int not null default 0,
+  role tinyint not null,
+  status tinyint not null default 0,
   primary key (id),
   unique (email),
   unique (phone)
